refactor: extract project modals
